what's the point vent room
if you also used it to expose bad opinion and then use it to hate ones?

Isn't it supposed to make us understand each other?

AI can never have soul. But better be, than would be and too sensitive

warn only, to not do anything would that be. If had be, then ensure it's just for them, and understand, your stuff won't last forever.

I am afraid, instead of seeing that the bridge collapsed, it's not. Just the police went karen, corruption. Licenses revoked for no reason and must pay again to make again from scratch.